Add flags for choosing algorithm in ecmult_multi benchmark
